Start moving pan view search code to its own module
Next step is to pull the construction of the Search UI into the module
as well.  Ideally, pan-view.c should just instantiate a Search thing,
without having to worry about exactly how it's constructed or destructed.
