+++ src/pan-view/Makefile.am
@@ -13,4 +13,7 @@ module_pan_view = \
 	%D%/pan-util.c	\
 	%D%/pan-util.h	\
 	%D%/pan-view.c	\
-	%D%/pan-view.h
+	%D%/pan-view.h	\
+	%D%/pan-view-search.c	\
+	%D%/pan-view-search.h
+
